Rename files to proper casing
