+++ client/extensions/woocommerce/app/settings/payments/payments-location-currency.js
