+++ client/extensions/woocommerce/app/settings/payments/payments-off-site.js
