+++ client/extensions/woocommerce/app/settings/payments/payments-on-site.js
